+++ src/services/jsonrpc.service.js
@@ -25,6 +25,11 @@ class JSONRPCService extends BaseService {
      * @type {Map<string, EventEmitter>}
      */
     this.requests = new Map()
+    /**
+     * 可接受的方法
+     * @type {Map<string, Function|Boolean>}
+     */
+    this.acceptMethods = new Map()
   }
   /**
    * 销毁
@@ -57,6 +62,10 @@ class JSONRPCService extends BaseService {
     // 定义ws server
     const host = opts.host || '0.0.0.0'
     const port = opts.port || 7897
+    Object.defineProperties(this, {
+      host: { value: host, writable: false, enumerable: true },
+      port: { value: port, writable: false, enumerable: true }
+    })
     this.wss = new WebSocket.Server({
       host,
       port,
@@ -116,19 +125,56 @@ class JSONRPCService extends BaseService {
 
   /**
    * 设置可接受rpc方法
-   * @param {string[]} acceptMethods
+   * @param {string[]|string} acceptMethods
    */
   setAcceptMethods (acceptMethods) {
     // 定义acceptMethods
     let methods = []
     if (acceptMethods) {
       if (typeof acceptMethods === 'string') {
-        methods = acceptMethods.split(',')
+        methods = acceptMethods.split(',').map(method => ({ name: method }))
       } else if (Array.isArray(acceptMethods)) {
-        methods = acceptMethods
+        methods = acceptMethods.map(method => {
+          if (typeof method === 'string') {
+            return { name: method }
+          } else if (typeof method.name === 'string') {
+            return { name: method.name, func: typeof method.func === 'function' ? method.func : undefined }
+          } else {
+            return null
+          }
+        }).filter(m => !!m)
       }
     }
-    this.acceptMethods = methods
+    // 设置acceptMethods
+    methods.forEach(method => {
+      const methodName = _.kebabCase(method.name)
+      this.acceptMethods.set(methodName, method)
+    })
+  }
+
+  /**
+   * 添加新的可接受方法
+   * @param {string} methodName 方法名
+   * @param {Function|undefined} methodFunc
+   */
+  addAcceptableMethod (methodName, methodFunc) {
+    methodName = _.kebabCase(methodName)
+    const methodData = { name: methodName }
+    if (typeof methodFunc === 'function') {
+      methodData.func = methodFunc
+    }
+    this.acceptMethods.set(methodName, methodData)
+  }
+
+  /**
+   * 移除新是可接受方法
+   * @param {string} methodName 方法名
+   */
+  removeAcceptableMethod (methodName) {
+    methodName = _.kebabCase(methodName)
+    if (this.acceptMethods.has(methodName)) {
+      this.acceptMethods.delete(methodName)
+    }
   }
 
   /**
@@ -147,7 +193,7 @@ class JSONRPCService extends BaseService {
       params: args,
       jsonrpc: '2.0'
     }
-    logger.tag(`Request:${methodName}`).log(`id=${reqData.id}`)
+    logger.tag(`Request:${methodName}`).debug(`id=${reqData.id}`)
     const emitter = new EventEmitter()
     this.requests.set(reqData.id, emitter)
     let objToSend = reqData
@@ -234,12 +280,18 @@ class JSONRPCService extends BaseService {
       let res = { jsonrpc: '2.0' }
       // 检测方法名是否可用
       const methodName = _.kebabCase(jsonRequest.method)
-      if (this.acceptMethods.indexOf(methodName) === -1) {
+      if (!this.acceptMethods.has(methodName)) {
         res.error = { code: 404, message: 'Method not found.' }
       } else {
+        const methodData = this.acceptMethods.get(methodName)
         // 进行本地调用
         try {
-          res.result = await jp.invokeMethod(methodName, jp.env.param, jsonRequest.params || {}, ws)
+          const params = jsonRequest.params || {}
+          if (!methodData.func) {
+            res.result = await jp.invokeMethod(methodName, jp.env.param, params, ws)
+          } else {
+            res.result = await methodData.func(params, ws)
+          }
         } catch (err) {
           let code = err.code
           let message = err.message
@@ -259,7 +311,7 @@ class JSONRPCService extends BaseService {
           }
         })
       }
-      logger.tag(`Invoked:${methodName}`).log(`id=${jsonRequest.id}`)
+      logger.tag(`Invoked:${methodName}`).debug(`id=${jsonRequest.id}`)
       return
     }
     // 回调类型判断
@@ -271,7 +323,7 @@ class JSONRPCService extends BaseService {
         logger.tag('Result').warn(`unknown id ${jsonResponse.id}`)
         return
       }
-      logger.tag(`Result`).log(`id=${jsonResponse.id}` + (!jsonResponse.error ? '' : `,code=${jsonResponse.error.code},message=${jsonResponse.error.message}`))
+      logger.tag(`Result`).debug(`id=${jsonResponse.id}` + (!jsonResponse.error ? '' : `,code=${jsonResponse.error.code},message=${jsonResponse.error.message}`))
       if (jsonResponse.result !== undefined) {
         emiter.emit('response', jsonResponse.result)
       } else if (jsonResponse.error !== undefined) {