+++ src/utils/config/index.d.ts
@@ -1,52 +1,30 @@
 import {
-  WalletSourceType,
-  WalletSourceConfig,
-  WalletCoinInfo,
-  TokenConfigBasic,
-  TokenConfigJadepool,
   ChainConfig,
   TokenConfig
 } from "../../models";
 
 /**
- * 根据config.configMods.{type} 将对象进行unset
- * @param types 逗号分隔类型
- * @param jsonObj 处理的对象
- */
-export function applyIgnoreKeys(types: string, jsonObj: object): object
-/**
- * 快速获取内存中的coin信息
- * @param chainKey
- * @param tokenNameOrNameOrAssetIdOrContract
- */
-export function fetchCoinCfgById(chainKey: string, tokenNameOrNameOrAssetIdOrContract: string): TokenConfigBasic
-/**
- * 获取区块链配置,仅有效链可获取
- * @param nameOrKeyOrCoreType
- */
-export function fetchChainCfg(nameOrKeyOrCoreType: string): ChainConfig
-/**
- * 获取该进程负责的全部链名称
- */
-export function fetchAllChainNames(): string[]
-/**
  * 读取默认配置中的token配置信息
+ * @deprecated
  * @param chainKey
  * @param coinName
  */
 export function loadCoinCfg(chain: string | ChainConfig, coinName: string): Promise<TokenConfig>
 /**
  * 获取实时的可用coinNames
+ * @deprecated
  * @param chainKey
  */
 export function loadAllCoinNames(chain: string | ChainConfig, includeDisabled?: boolean): Promise<string[]>
 /**
  * 获取实时的区块链配置
+ * @deprecated
  * @param chainKey
  */
 export function loadChainCfg(chainKey: string): Promise<ChainConfig>
 /**
  * 获取实时的全部链名称
+ * @deprecated
  */
 export function loadAllChainNames(includeDisabled?: boolean): Promise<string[]>
 
@@ -56,9 +34,29 @@ export function loadAllChainNames(includeDisabled?: boolean): Promise<string[]>
  * @param coinName
  * @deprecated
  */
-export function fetchCoinCfg(coinName: string): any
+export function fetchCoinCfg(): void;
 /**
  * 该函数已无法使用
  * @deprecated
  */
-export function fetchAllCoinNames(chainKey?: string): string[]
+export function fetchAllCoinNames(): void;
+/**
+ * 该函数已无法使用
+ * @deprecated
+ */
+export function applyIgnoreKeys(): void;
+/**
+ * 该函数已无法使用
+ * @deprecated
+ */
+export function fetchCoinCfgById(): void;
+/**
+ * 该函数已无法使用
+ * @deprecated
+ */
+export function fetchChainCfg(): void;
+/**
+ * 该函数已无法使用
+ * @deprecated
+ */
+export function fetchAllChainNames(): void;