+++ src/utils/config/index.js
@@ -1,53 +1,6 @@
-const _ = require('lodash')
 const NBError = require('../../NBError')
 const consts = require('../../consts')
 const jp = require('../../jadepool')
-
-const applyIgnoreKeys = (types, jsonObj) => {
-  let typeArr = types.split(',')
-  typeArr = _.intersection(_.keys(jp.config.configMods), typeArr)
-  typeArr.forEach(typeKey => {
-    // 移除ignores
-    _.forEach(jp.config.configMods[typeKey], key => _.unset(jsonObj, key))
-  })
-  return jsonObj
-}
-
-const fetchChainCfg = (nameOrKeyOrCoreType) => {
-  let chainCfg = _.find(jp.config.chain, { key: nameOrKeyOrCoreType })
-  if (!chainCfg) {
-    chainCfg = _.find(jp.config.chain, { Chain: nameOrKeyOrCoreType })
-  }
-  if (!chainCfg) {
-    chainCfg = _.find(jp.config.chain, { CoreType: nameOrKeyOrCoreType })
-  }
-  return chainCfg
-}
-
-const fetchAllChainNames = () => {
-  if (jp.env.server === consts.SERVER_TYPES.MAIN && jp.env.processType === consts.PROCESS.TYPES.BLOCKCHAIN) {
-    return jp.env.param ? [ jp.env.param ] : []
-  } else {
-    return _.keys(jp.config.chain)
-  }
-}
-
-/**
- * 根据denom获取CoinCfg
- * @param {string} denomOrName
- * @returns {{ name: string, Rate: number, TokenName: string }}
- */
-const fetchCoinCfgById = (chainKey, tokenNameOrAssetIdOrContract) => {
-  const cfgs = jp.config.coin[chainKey]
-  if (!cfgs || !tokenNameOrAssetIdOrContract) return null
-  return _.find(cfgs, cfg => {
-    return cfg.name === tokenNameOrAssetIdOrContract ||
-      cfg.TokenName === tokenNameOrAssetIdOrContract ||
-      cfg.Contract === tokenNameOrAssetIdOrContract ||
-      cfg.contract === tokenNameOrAssetIdOrContract ||
-      cfg.assetId === tokenNameOrAssetIdOrContract
-  })
-}
 
 const loadCoinCfg = async (chainKey, coinName) => {
   const ConfigDat = jp.getModel(consts.MODEL_NAMES.CONFIG_DATA)
@@ -133,20 +86,20 @@ const loadAllCoinNames = async (chainKey, includeDisabled = false) => {
 }
 
 // ------------------------ 已废弃方法 ------------------------
-const fetchCoinCfg = () => { throw new NBError(10001, `unsupported method[fetchCoinCfg]`) }
-const fetchAllCoinNames = () => { throw new NBError(10001, `unsupported method[fetchAllCoinNames]`) }
+const deprecatedMethod = function (methodName) {
+  return () => { throw new NBError(10001, `unsupported method[${methodName}]`) }
+}
 
 module.exports = {
-  applyIgnoreKeys,
-  fetchCoinCfgById,
-  fetchChainCfg,
-  fetchAllChainNames,
   // 实时读取
   loadCoinCfg,
   loadChainCfg,
   loadAllCoinNames,
   loadAllChainNames,
   // 废弃方法
-  fetchCoinCfg,
-  fetchAllCoinNames
+  fetchCoinCfg: deprecatedMethod('fetchCoinCfg'),
+  fetchAllCoinNames: deprecatedMethod('fetchAllCoinNames'),
+  fetchCoinCfgById: deprecatedMethod('fetchCoinCfgById'),
+  fetchChainCfg: deprecatedMethod('fetchChainCfg'),
+  fetchAllChainNames: deprecatedMethod('fetchAllChainNames')
 }