+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Logs
 
+## V0.12.0
+
+Features:
+
+* 新增config.service服务，全权处理全部的配置读取与写入事宜
+* utils.configSetuper下的所有方法设置为deprecated，同时将不再往config对象中写入内容，仅提供初始化功能
+* 移除utils.config.fetch**系列方法
+* 移除utils.config.applyIgnoreKeys方法
+
 ## V0.11.18
 
 Fixes: